Split TiffWriter to the specialized classes
